refactor: resolve database driver from circle.yml

+++ test/index.js
@@ -7,7 +7,7 @@ import tryCatch from '../src/utils/try-catch';
 
 import { getTestApp } from './utils/get-test-app';
 
-const { env: { APPVEYOR, CIRCLECI, CIRCLE_NODE_INDEX } } = process;
+const { env: { APPVEYOR, CIRCLECI } } = process;
 
 before(function (done) {
   this.timeout(120000);
@@ -16,39 +16,9 @@ before(function (done) {
 
   tryCatch(async () => {
     const path = resolvePath(__dirname, 'test-app');
+    const execOpts = { cwd: path };
 
-    const execOpts = {
-      cwd: path,
-      env: {
-        ...process.env
-      }
-    };
-
-    if (APPVEYOR) {
-      // Do nothing for now.
-    } else if (CIRCLECI) {
-      let driver;
-
-      switch (CIRCLE_NODE_INDEX) {
-        case '0':
-          driver = 'pg';
-          break;
-
-        case '1':
-          driver = 'mysql2';
-          break;
-
-        default:
-          driver = 'sqlite3';
-          await exec('lux db:reset', execOpts);
-          break;
-      }
-
-      process.env.DATABASE_DRIVER = driver;
-      execOpts.env.DATABASE_DRIVER = driver;
-
-      console.log(`Starting test suite using database driver '${driver}'.`);
-    } else {
+    if (!APPVEYOR && !CIRCLECI) {
       await exec('lux db:reset', execOpts);
     }
